+++ packages/croquet/teatime/src/webrtc.js
@@ -1,3 +1,4 @@
+import { App } from "./_HTML_MODULE_"; // eslint-disable-line import/no-unresolved
 import urlOptions from "./_URLOPTIONS_MODULE_"; // eslint-disable-line import/no-unresolved
 
 const ICE_NEGOTIATION_MAX = 15_000; // maximum ms between sending our offer and the data channel being operational.  analogous to the controller's JOIN_FAILED_DELAY, between sending of JOIN and receipt of SYNC.
@@ -21,7 +22,7 @@ function getRandomString(length) {
 }
 
 export class CroquetWebRTCConnection {
-    static get name() { return 'CroquetWebRTCConnection'; }
+    static get name() { return `${App.libName}WebRTCConnection`; }
 
     constructor(registryURL) {
         this.clientId = getRandomString(4); // used locally to tag console messages for debug purposes; gets appended with a numeric id from SessionRunner on connection
@@ -152,11 +153,10 @@ export class CroquetWebRTCConnection {
                 const msgData = JSON.parse(rawMsg.data);
                 if (msgData.what === 'READY') {
                     // sent by the SessionRunner to indicate socket acceptance.
-                    // now includes the SessionRunner-assigned client id.
+                    // now includes the SessionRunner-assigned client id, and
+                    // an array of ICE servers.
                     this.clientId += `_${msgData.id}`;
-                    if (msgData.iceServers && !globalThis.iceServersP) {
-                        globalThis.iceServersP = Promise.resolve(msgData.iceServers);
-                    }
+                    this.iceServers = msgData.iceServers;
                     resolve();
                     return;
                 }
@@ -325,42 +325,7 @@ export class CroquetWebRTCConnection {
     }
 
     async createPeerConnection() {
-        if (!globalThis.iceServersP) {
-            globalThis.iceServersP = new Promise((resolve, reject) => {
-                // fetch STUN and TURN details from Open Relay (https://www.metered.ca/tools/openrelay/)
-                let tries = 3;
-                const fetchServers = () => {
-                    fetch(process.env.ICE_SERVERS_URL)
-                    .then(response => {
-                        if (response.status !== 200) throw Error(`status=${response.status}`);
-                        return response.json();
-                    }).then(serversObj => {
-                        // an opportunity to edit the supplied servers - for example, if we want to add an alternative TURN service.
-                        // [
-                        //   { urls: "stun:stun.l.google.com:19302" },
-                        //   { urls: "stun:stun.l.google.com:5349" },
-                        // ].forEach(stun => serversObj.unshift(stun));
-                        // console.log(serversObj);
-                        resolve(serversObj);
-                    }).catch(err => {
-                        console.warn("error in fetching ICE servers", err);
-                        if (--tries > 0) setTimeout(fetchServers, 250);
-                        else reject(); // caught below, throwing a 4200 (fatal) error
-                    });
-                };
-                fetchServers();
-            });
-        }
-
-        let iceServers;
-        try {
-            iceServers = await globalThis.iceServersP;
-        } catch (err) {
-            const error = new Error("failed to fetch ICE servers");
-            error.code = 4200; // do not retry
-            throw error;
-        }
-
+        const { iceServers } = this;
         const iceReportedErrors = new Set(); // do our own filtering
         const pc = this.pc = new globalThis.RTCPeerConnection({ iceServers });
         pc.onnegotiationneeded = _e => {